Rename projection.js to projectionUtils.js to better match its purpose
